+++ testsrequire.py
@@ -1,6 +1,7 @@
 
 tests_require = [
     'pytest',
+    'pytest-asyncio',
     'sqlalchemy>=2.0.0,<2.1.0',
     'greenlet>=2.0.1',
     'alembic',